Avoid copy_from_slice() in marshal_X()
In working towards panic-free generated code, avoid the usage of
copy_from_slice() from the marshal_X() helpers, with X any of the
native integer types.

Instead of converting the source array into a slice and copying it into the
destination slice, cast the destination slice to an array (reference) of
appropriate length and copy into that by means of a plain assignment.

Signed-off-by: Nicolai Stange <nstange@suse.de>