Make ::into_bufs_owner() an associated function
Currently, ::into_bufs_owner() is a member method. However,
various instances thereof take a Box<Self> and if Box<> is not
coming from standard Rust, it's disallowed for 'self'.

In preparation of using the allocators-api2 crate, and importing the
Box<> implementation therefrom, make ::into_bufs_owner() an associated
function throughout.

Signed-off-by: Nicolai Stange <nstange@suse.de>