Don't wrap a TpmBuffer's owned Vec<> in a Zeroizing
If the "zeroize" feature is enabled, the generated TpmBuffer::Owned
Vec gets wrapped in a Zeroizing<> to enforce zeroization at Drop.

However, if the Vec<> is not coming coming from the standard Rust,
but e.g. from the transitional allocator-api2 crate, Vec<> would
not implement the prerequisite Zeroize trait.

To work around this, do not wrap a TpmBuffer::Owned's Vec<>
in a Zeroizing, but implement Drop explicitly for TpmBuffer
and do the zeroization manually from there.

Signed-off-by: Nicolai Stange <nstange@suse.de>