Provide marshal_bytes() helper in generated code and use it throughout
When marshalling a member of type array of bytes, the contents are
getting copied directly to the output buffer by means of a
[u8]::copy_from_slice(). However, that is not panic-free, and
a panic-free alternative would require 'unsafe {}' code.

In order to provide a maintainable panic-free alternative, provide
a marshal_bytes() helper and use it throughout when marshalling
an array of bytes. Select between either of the two possible
implementations based on whether the --enable-unsafe-panic-free
CLI flag has been specified by the user.