Handle possibly empty enums in generated code
In case where all of a generated enum's variants have some cfg() deps
associated with them, and all such are disabled, the enum will be
empty.

This by itself is not a problem, Rust allows empty enums, they just
cannot get instantiated. However, certain constructs are not allowed,
most notably a #[repr(uN)] attribute, or a match on a reference to
the enum with no wildcard match arm. Make provisions such that the
generated code still compiles fine.

Signed-off-by: Nicolai Stange <nstange@suse.de>